requirements.txt
# ============================================================
# DeskMate Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# AI API
google-generativeai>=0.4.0

# Screen Capture
mss>=9.0.0

# Image Processing
Pillow>=10.0.0

# UI Framework
customtkinter>=5.2.0

# System Tray Icon
pystray>=0.19.0

# Global Hotkeys
keyboard>=0.13.5

# Window Info (Windows)
PyGetWindow>=0.0.9

# Process Info
psutil>=5.9.0

# Config
PyYAML>=6.0

# Clipboard
pyperclip>=1.8.0

# HTTP Requests (live docs fetcher)
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0
